+++ apps/migration-plan/src/app/pages/Metricas/Metricas.tsx
@@ -143,3 +143,4 @@ const Metricas: FC = () => {
 export default Metricas;
 
 
+