+++ apps/migration-plan/src/app/pages/Riesgos/Riesgos.tsx
@@ -96,3 +96,4 @@ const Riesgos: FC = () => {
 export default Riesgos;
 
 
+